requirements.txt
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
python-dotenv==1.0.1
requests==2.32.3
serpapi==0.1.5
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
